send/receive complete lines, fix for Beast 1/n-1 record splitting
The Beast CVE fix causes only one octet to be sent in the first socket
packet and then the remaining data is sent normally, this is known as
1/n-1 record splitting. The example and test SSL code sent short
messages and then did a sock.recv(1024). We had always received the
entire message in one sock.recv() call because it was so short. But
sock.recv() does not guarantee how much data will be received, thus
this was a coding mistake. The solution is straight forward, use
newlines as a record separator and call sock.readline() instead of
sock.recv(). sock.readline() calls sock.recv() internally until a
complete line is read or the socket is closed.