Replace epydoc documentation generator with Sphinx autodoc
Epydoc is deprecated and no longer maintained, we need to switch to
another Python API documentation generator that can operate on
embedded doc strings. At this moment Sphinx is the preferred
documentation tool for Python and with it's autodoc extension can
build doc from the embedded docstrings, as such Sphinx seems like the
best replacement for epydoc.